[MOV] im_livechat: move Timers outside of public_livechat.js
Task-2212347
